Add steps to update Insights container images
Adding steps to update the Insights container images before the upgrade, without which installer will fail.

The connected guide just requires one step, which is added directly to the prerequisites.
The disconnected guide needs multiple steps. Adding an assembly for this with the two ways this can be done as two proc modules.