Rework navigation structure to simplify it (#4068)
* Remove unused code

* Rework navigation structure

This creates landing pages for each guide where all build flavors are
shown. By default they are collapsed because they're fairly long. Then a
new individual landing page is created for every build flavor.

Then the navigation is changed so there is a link to the landing page.
The version switcher remains in place to allow users to switch between
different versions of the guide.

* Convert remaining .json files to the new structure

* Update main page to drop reference to flavor in navigation

---------

Co-authored-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>