Fix example URL for multiple CV envs (#4101)
This patch moves the attributes for the examples in "Content view
environment ordering and priority" from the assembly to the module. This
is required to show the correct example URL for the second example that
reorderes the content view environments on a host.

Refs PR #3609

(cherry picked from commit 8c54b586ed7707582b47765da8f411b97314aad0)