Use name and major version as description of OS entry
If you create an operating system entry by using Hammer CLI & copy/paste
the example from the docs, you end up with "My_Operating_System" on the
Hosts > Operating Systems page. The reason is that the page in the WebUI
shows the description in the "Title" column.

This patch ensures that it aligns with other OS entries.