Drop run the following command
"Run the following command" and similar phrases do not add any value. A
step in the docs following by a code block _always_ mean that you have
to run a command to follow the procedure.

If you run "hammer proxy", you'll need to provide the ID of your Smart
Proxy. It does not depend on the host on which your run your Hammer CLI
commands on.

$ hammer proxy info
Missing arguments for '--id'.
$ hammer capsule info
Missing arguments for '--id'.

Fixes #3122