Rename hammer-cli to cli in file names (#4406)
We have a convention to use "by-using-cli" or "by-using-web-ui" in
anchors. Therefore, it makes maintenance easier if we also apply this to
module names that document procedures for Hammer CLI in constrast to
WebUI or API.

The main benefit is that you can search for the anchor and also find the
assembly that includes the modules.

Fixes #4350

(cherry picked from commit 56c7fd9c36ab7e178ab9bba11fde44efc6f711d9)