Address asciidoctor-dita-vale for administering hosts (#4416)
* Split multiple modules into web UI and CLI

* Add abstract role and create basic abstracts

* Drop a module on creating hosts

This saves us from having to come up with an abstract for a short module
that doesn't provide enough value on its own.

It involves deleting some xrefs that pointed to the module. I think it
should be fine because really, the module didn't contain anything
substantial anyway.

* Drop a module about switching hosts

This saves us from the need to write an abstract for a very short module
that doesn't stand that well on its own.

Additionally, it seems to fit quite well into the module on browsing
hosts.

* Drop a module about searching lifecycle status

Another one I'd prefer to avoid having to write an abstract for.

* Minor fixes

Co-authored-by: Maximilian Kolb <mail@maximilian-kolb.de>

---------

Co-authored-by: Maximilian Kolb <mail@maximilian-kolb.de>